document_processing_service/requirements.txt
gensim==4.2.0
PyPDF2==3.0.0
python-docx==0.8.11
chromadb==0.2.1
